Update tmp
Closes #173


Hi.

I see you are a premium user. And problem in installing module. 

Try to clear data and activate premium again 👍

If this not works then send me screenshots or video recording of actual problem

Or wait for next update so i can fix it for you

You can also contact on telegram for faster response : @AraafRoyall

Or reply here...